src/third_party/blink/web_tests/fast/css-grid-layout/percent-intrinsic-track-breadth-expected.txt

CONSOLE WARNING: line 48: Percentages row tracks and gutters for indefinite height grid containers will be resolved against the intrinsic height instead of being treated as auto and zero respectively. This change will happen in M70, around October 2018. See https://www.chromestatus.com/feature/6708326821789696 for more details.
This test checks that percentage track breadths of intrinsic size are treated as auto.

XXXXX
PASS
